# [Graphbolt] Add `cat` optimization for UniformPick