## Description
<!-- Brief description. Refer to the related issues if existed.
It'll be great if relevant reviewers can be assigned as well.-->
We've optimized the `SampleNeighborsImpl` function in `csc_sampling_graph.cc`. This PR primarily resolves this [Issue#5817 Optimize CSC neighbor sampler performance.](https://github.com/dmlc/dgl/issues/5817)

The original algorithm, when conducting multithreaded Neighbor Sampling, would place all sampled results into a general vector of Tensor. We want to optimize this part.

### Problem Analysis

Let's consider a scenario where we have $n$ `Tensors`(same size) and $m$ `threads`, and we need to concatenate all `Tensors` into one large `Tensor`. We consider two different strategies:

- Strategy **A**: Each `thread` uses `torch::cat` to combine up to $\frac{n}{m}$ `Tensors` into a `big Tensor`, then `torch::cat` is used again to combine the $m$ `big Tensors` into the `result`.
- Strategy **B**: Each `thread` only returns a `std::vector<Tensor>`, and in the main program, the results from each thread are combined into a `std::vector<Tensor> all_tensors`. Finally, `torch::cat` is used to combine the `all_tensors` with $n$ elements into the `result`. (currently strategy)

A necessary condition for Strategy **A** to outperform Strategy **B** is:

- **[Assumption 1]**: The performance of `torch::cat` when merging **fewer large-Tensors** is superior to that of merging **many small-Tensors**. We conducted a simple test to verify this: (where `small` represents a group of small tensors and `big` represents a group of large tensors)

| `small_num` | `small_size` | `big_num` | `big_size` | `device` |  `small`  |   `big`   |     `gap`     |
| :---------: | :----------: | :-------: | :--------: | :------: | :-------: | :-------: | :-----------: |
|   $10000$   |    $2000$    |   $20$    | $1000000$  |  `CPU`   | $1.1438s$ | $0.3369s$ |  $3.4\times$  |
|   $10000$   |    $2000$    |   $200$   |  $100000$  |  `CPU`   | $1.1321s$ | $0.3941s$ |  $2.9\times$  |
|   $10000$   |    $2000$    |   $20$    | $1000000$  |  `GPU`   | $0.1146s$ | $0.0020s$ | $57.3\times$  |
|   $10000$   |    $2000$    |   $200$   |  $100000$  |  `GPU`   | $0.1133s$ | $0.0037s$ | $30.62\times$ |

We found **[Assumption 1]** to be true, so we proceeded with the next optimization step.

Based on our performance tests, we observed that **if the results obtained by each thread are first stored within the thread and `cat` is performed once within the thread to obtain a smaller Tensor, and finally `cat` is performed externally on the Tensors obtained from all threads, the efficiency will be higher.** In other words, Strategy A is superior. The detailed performance test experiment is as follows: 

### Time Performance Test

![exp_cat](https://github.com/dmlc/dgl/assets/103916249/87e63bb6-b4d8-47f8-8ad7-722e323c634d)

![exp_cat_time_table](https://github.com/dmlc/dgl/assets/103916249/39486799-5846-4ab3-b271-a53240dc6808)

We found that except in very few cases, this optimization could bring about a performance improvement of around $2$ times. **For red cells in Graph3**, when the data is too large, it is possible that multiple `cat` operations within a thread may lead to performance degradation.

### Memory Usage Test

We performed memory tests using `Massif` in `Valgrind`, and for the sake of test accuracy, the entire test was conducted in a C++ environment, looping each test 5 times. (Massif is a tool in Valgrind, **used to measure the amount of heap memory used by an application**)

![exp_cat_memo](https://github.com/dmlc/dgl/assets/103916249/eb48e52b-6f88-4f1d-8e30-c5e34f377c9c)

![exp_cat_memo_table](https://github.com/dmlc/dgl/assets/103916249/fa878cb7-170c-4468-bb6e-731e4516aca9)

We only performed memory tests on two commonly used graphs. **For Graph1**, there was also an improvement in terms of memory. **For Graph2**, since there were only $10$ sampled nodes, our strategy did not significantly impact it. 

Therefore, this optimization is effective.