@frozenbugs 2023-07-20
What does neighbors_per_etype mean?