@keli-wen 2023-07-21
> What does neighbors_per_etype mean?

According to the generate code:

```py
def random_graph_with_fixed_neighbors(num_nodes,
                                      neighbors_per_etype,
                                      num_ntypes,
                                      num_etypes):
    neighbors_per_node = num_etypes * neighbors_per_etype
    csc_indptr = torch.arange((num_nodes+1)) * neighbors_per_node
    num_edges = num_nodes * neighbors_per_node
    indices = torch.randint(0, num_nodes, (num_edges,))
    type_per_edge = torch.arange(num_etypes).repeat_interleave(
        neighbors_per_etype
    ).repeat(num_nodes)

    node_type_offset = torch.sort(
        torch.randint(0, num_nodes, (num_ntypes + 1,))
    )[0]
    node_type_offset[0] = 0
    node_type_offset[-1] = num_nodes

    # Add the metadata to the graph.
    ntypes = { f"n{i}" : i for i in range(num_ntypes) }
    etypes = {
        (f"n{(i % num_ntypes) + 1}", f"e{i}", f"n{(i // num_etypes) + 1}") : i
            for i in range(num_etypes)
    }
    metadata = gb.GraphMetadata(ntypes, etypes)

    return gb.from_csc(csc_indptr, indices, node_type_offset, type_per_edge, None, metadata)
```

`neighbors_per_etype` denote the number of neighbors of each edge type of the node.