@mfbalin 2023-07-25
So the speedup this PR is achieving over the existing implementation may be due to the possibility that `torch.cat` doesn't have a parallel implementation. Profiling the code would reveal whether this is the case. Otherwise, doing two levels of `torch.cat` operations only increases the work. Each edge is concatenated two times.