@peizhou001 2023-07-25
> 

That's true, and the speedup is not very much as you can see in the diagram. According to our test, it performs better when the data volume is small, maybe because cat huge tensors can leverage bulk copy or vector instruction (just suspect)? so it is a tradeoff to decide whether merge it. For rollout cat, it could be a backup improvement for future.