@frozenbugs 2023-07-25
Hi @mfbalin, thanks for your suggestion, we will further optimize the performance by allocating the memory first to avoid concatenation. We will merge this PR for now since there are some measurable improvement.