@peizhou001 2023-07-26
> neighbors

It can also be calculated when `probs_or_mask` exists, I have ever implmented it in #5717 .